requirements.txt
emoji==0.4.5
h5py==2.7.0
Keras==2.0.9
numpy==1.13.3
scikit-learn==0.19.0
tensorflow==1.8.0
text-unidecode==1.0
imgpy
tqdm
natsort
Pillow

